LoadSpectSeq: use an fread API to read wxInputStream data
This creates an fread compatible API on wxInputStream objects and
uses that API in LoadSpectSeq. This is to help the transition from
the wxWidgets wxInputStream API to the C FILE API.

+++ src/spect.h
@@ -112,7 +112,7 @@ double GetFrameRms(SpectFrame *frame, int amp);
 struct SpectSeq
 {
 	int  numframes;
-	int  amplitude;
+	short amplitude;
 	int  spare;
 	char *name;
 
@@ -124,7 +124,7 @@ struct SpectSeq
 	int  grid;
 	int  bass_reduction;
 	int max_x;
-	int max_y;
+	short max_y;
 	int file_format;
 };
